dashboard/requirements.txt
pandas==1.3.3
numpy==1.21.2
lxml==5.2.1
streamlit==1.34.0
python-dotenv==1.0.1
st-annotated-text==4.0.1
seaborn==0.13.2
plotly==5.22.0
mlflow==2.13.0